Make sync_gitlab_issues more performant by using an iterator
Previously, gl.issues.list was fetching all issues before
processing them. That lead to a medium sized gap where nothing
happened in the shell because the command was waiting for the issues
to be fetched. This commit converts into using a paged iterator for
the issues, and the iretator can be processed with tqdm instantly so
there's no "downtime" anymore for waiting the issues. The old progress
bar is also retained.